fixup! 1. `ERT_Main` and `WaterContentMain` link with the UI  2. changed folder name Data Inversion into data_inversion for referencing 3. change file selection from designated location to user selection.
