+++ data_inversion/inversion_result_.png
